Fix initial loading bug
Because the repo info query uses the enabled flag (which was a late addition)
the query is initially loading in that state, which is technically true.
We should detect that with both the initial loading flag and our own
repo url state to allow first time users using this.

More info on react-query enabled and initial loading here:
https://tanstack.com/query/v4/docs/guides/disabling-queries#isinitialloading